requirements.txt
anyio==3.6.1
ariadne==0.15.1
autopep8==1.6.0
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
graphql-core==3.2.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.20.1
toml==0.10.2
typing-extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
